+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ## New in 2.1.1
 * adjusted the build script to publish only dated releases
+* fixed bug about enumerable-assertion has-count only expecting dojos (#60)
 
 ## New in 2.1.0 (Released 2017/11/15)
 * extended the has-any enumerable-assertion to support multiple assertions (#59)